% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcontrol.R
\name{calculate_qc_scores}
\alias{calculate_qc_scores}
\title{Calculate QC scores from sequencing}
\usage{
calculate_qc_scores(patients_dir, trimmed)
}
\arguments{
\item{patients_dir}{data from module of FastQC output}

\item{trimmed}{it can be T or F}
}
\description{
Prepare the data for plotting
}
