% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimpleReport.R
\name{caratula}
\alias{caratula}
\title{Generate Results' Report - Section Caratula}
\usage{
caratula(patient_dir, path_metadata)
}
\arguments{
\item{path_metadata}{path that indicates were the Metadata excel is located}

\item{patients_dir}{indicated the patient directory that will be analyzed.}
}
\description{
It generates the report in pdf format that register all the predictions and the analizes of the microbiome composition of a determined patient.
}
