% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHuman.R
\name{count_Vias}
\alias{count_Vias}
\title{Count Vias}
\usage{
count_Vias(patients_dir, de_host)
}
\arguments{
\item{patients_dir}{path to the folder of the patients that will be analyzed.}

\item{de_host}{indicates which alignment-based de-hosting software will be used.}
}
\description{
Generates a df that counts the number of identified pathways for each sample.
}
