% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_Tax.R
\name{counts_Tax}
\alias{counts_Tax}
\title{Counts Taxonomy}
\usage{
counts_Tax(patients_dir, source, de_host, conEukaryota)
}
\arguments{
\item{source}{kjsnfn}

\item{patient_dir}{is a dataframe fskbfjsf}
}
\value{
cants_original dataframe that contains how many subjects from the same taxonomic level were found.
}
\description{
kvjdfnkvjdf
}
\examples{
counts_taxonomy <- counts_Tax(otus_table_kraken04)
}
