% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcontrol.R, R/generateQCreport.R
\name{create_FQCdata}
\alias{create_FQCdata}
\title{create FastQC data}
\usage{
create_FQCdata(dat)

create_FQCdata(dat)
}
\arguments{
\item{dat}{data from module of FastQC output}
}
\description{
Prepare the data for plotting

Prepare the data for plotting
}
