% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcontrol.R
\name{deHosting}
\alias{deHosting}
\title{De Hosting}
\usage{
deHosting(patient_dir, de_host, compatible_DRAGEN = FALSE)
}
\arguments{
\item{patients_dir}{fskbfjsf}

\item{de_host.}{It can be "Bowtie", "BWA" or "RSubread".}
}
\value{
qc_scores is  dataframe that records qc mean and \%Q30
}
\description{
kvjdfnkvjdf
}
