% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDRAGEN.R
\name{download_DRAGENReport}
\alias{download_DRAGENReport}
\title{Download DRAGEN Report}
\usage{
download_DRAGENReport(patient_dir, bs_path, de_host)
}
\arguments{
\item{patient_dir}{fskbfjsf}

\item{bs_path}{path to bs}

\item{project_id}{it kvsbdkj}
}
\description{
kvjdfnkvjdf
}
\examples{
path_report <- download_DRAGENReport(patient_dir ="~/Biota/Nuevas13/Muestras/33" )
}
