% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadBowtie2.R
\name{downloadBowtie2}
\alias{downloadBowtie2}
\title{downloadBowtie2}
\usage{
downloadBowtie2()
}
\value{
The path where the ejecution file is
}
\description{
Downloads and decompresses the Bowtie2 software
}
