% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadBS.R
\name{downloadBS}
\alias{downloadBS}
\title{Download BS (Base Space)}
\usage{
downloadBS(dir_bs)
}
\arguments{
\item{dir_bs}{path where to store the executable Base Space}
}
\value{
path of the bs executable
}
\description{
downloads and installs the executable of bs for DRAGEN metagenomics processes by CLI Base Space aplication
}
