% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFastQC.R
\name{downloadFastQC}
\alias{downloadFastQC}
\title{downloadFastQC}
\usage{
downloadFastQC()
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the FastQC software
}
