% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSamples.R
\name{downloadFolderDrive}
\alias{downloadFolderDrive}
\title{Download a folder from Drive}
\usage{
downloadFolderDrive(folder_url, local_folder)
}
\arguments{
\item{folder_url}{is a url.}

\item{local_folder}{is the path where all the folders will be saved.}
}
\description{
Downloads samples from a folder in googledrive and saves them respecting the structure needed.
}
\examples{
downloadFolderDrive(folder_url = "https://drive.google.com/drive/folders/1TayWJGJK9KJHExisiTNyk_cM-fxDCR4n?usp=drive_link", local_folder = "/media/4tb2/Daniela/Biota/Muestras")
}
