% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSamtools.R
\name{downloadSamtools}
\alias{downloadSamtools}
\title{downloadSamtools}
\usage{
downloadSamtools()
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the Samtools software
}
