% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTrimGalore.R
\name{downloadTrimGalore}
\alias{downloadTrimGalore}
\title{downloadTrimGalore}
\usage{
downloadTrimGalore()
}
\value{
The path where the .exe file is located
}
\description{
Downloads and decompresses the TrimGalore software
}
