% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNames.R
\name{formatNames}
\alias{formatNames}
\title{Format names}
\usage{
formatNames(patients_dir)
}
\arguments{
\item{patient_dir}{path to the folder that stores all the samples}
}
\description{
It changes the format in which samples are named in order to match with dragen requirments
}
