% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateARcount.R
\name{generateARcounts}
\alias{generateARcounts}
\title{Generate AR counts}
\usage{
generateARcounts(patients_dir)
}
\arguments{
\item{patients_dir}{path to the folder that stores all the samples folders.}
}
\value{
list with: outD, out, ConteosKD.
}
\description{
Generates one excel file for each taxonomic level containing the relative abundances of all patients in patient_dir.
}
\examples{
counts_taxonomy <- counts_Tax(otus_table_kraken04)
}
