% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCompleteReport.R
\name{generateCompleteReport}
\alias{generateCompleteReport}
\title{Generate Complete Results' Report}
\usage{
generateCompleteReport(patient_dir)
}
\arguments{
\item{patients_dir}{indicated the patient directory that will be analyzed.}
}
\description{
It generates the report in pdf format that register all the predictions and the analizes of the microbiome composition of a determined patient.
}
