% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOTUsTable.R
\name{generateOTUsTable_Individual}
\alias{generateOTUsTable_Individual}
\title{Generate OTUs Table Individual}
\usage{
generateOTUsTable_Individual(
  patient_dir,
  source,
  conEukaryota = FALSE,
  de_host,
  paraBiota
)
}
\arguments{
\item{patient_dir}{fskbfjsf}

\item{source}{it must be "DRAGEN" or "KRAKEN"}
}
\value{
tabla_otus dataframe that formats KRAKEN or DRAGEN reports in a comparable way
}
\description{
kvjdfnkvjdf
}
\examples{
tabla_otus_kraken <- generateOTUsTable_Individual(patient_dir ="~/Biota/Nuevas13/Muestras/33", source= "KRAKEN" )
}
