% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOTUsTable.R
\name{generateOTUsTableGrupal}
\alias{generateOTUsTableGrupal}
\title{Generate OTUs Table Grupal}
\usage{
generateOTUsTableGrupal(
  patients_dir,
  source,
  conEukaryota = F,
  de_host,
  paraBiota
)
}
\arguments{
\item{patients_dir}{path to the folder that stores all the samples folders.}

\item{source}{it must be "DRAGEN" or "KRAKEN"}

\item{paraBiota}{it can be TRUE or FALSE. If set to TRUE transform Cutibacterium modestum}
}
\value{
marged_df dataframe that formats KRAKEN or DRAGEN reports in a comparable way. The columns contains the samples and the rows the OTUs
}
\description{
kvjdfnkvjdf
}
\examples{
tabla_otus_kraken <- generateOTUsTable_Grupal(patient_dir ="~/Biota/Nuevas13/Muestras", source= "KRAKEN" )
}
