% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHuman.R
\name{generatePathwayReport}
\alias{generatePathwayReport}
\title{Run HUMAnN}
\usage{
generatePathwayReport(id, de_host)
}
\arguments{
\item{de_host}{indicates which alignment-based de-hosting software will be used.}

\item{patient_dir}{path to the folder of the patient that will be analyzed.}
}
\description{
Executes concat of R1 and R2 fastq files, executes HUMAnN 3.0 and generates CPM.
}
