% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHuman.R
\name{generatePathwaysTable}
\alias{generatePathwaysTable}
\title{Generate Pathways Table}
\usage{
generatePathwaysTable(patients_dir, de_host)
}
\arguments{
\item{patients_dir}{path to the folder of the patients that will be analyzed.}

\item{de_host}{indicates which alignment-based de-hosting software will be used.}
}
\description{
Joins CPM tables, adds information of classes from Metacyc and calculates Relative Abundances.
}
