% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateScore.R
\name{generateScore}
\alias{generateScore}
\title{Generate Balance Score}
\usage{
generateScore(patient_dir)
}
\arguments{
\item{patients_dir}{indicated the patient directory that will be analyzed.}
}
\description{
It generates the report in pdf format that register all the predictions and the analizes of the microbiome composition of a determined patient.
}
