% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimpleReport.R
\name{generateSimpleReport}
\alias{generateSimpleReport}
\title{Generate Results' Report}
\usage{
generateSimpleReport(patient_dir, path_metadata)
}
\arguments{
\item{patients_dir}{indicated the patient directory that will be analyzed.}
}
\description{
It generates the report in pdf format that register all the predictions and the analizes of the microbiome composition of a determined patient.
}
