% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_TaxonomicLevels.R
\name{group_TaxonomicLevels}
\alias{group_TaxonomicLevels}
\title{Group Taxonomic Levels}
\usage{
group_TaxonomicLevels(
  patients_dir,
  tabla_otus,
  source,
  de_host,
  especies_seleccionadas = c(),
  subespecies_seleccionadas = c(),
  nombre_extra = "",
  conEukaryota
)
}
\arguments{
\item{patients_dir}{path to the directory were are stored all the patients analyzed}

\item{tabla_otus}{is a dataframe fskbfjsf}

\item{source}{it must be "DRAGEN" or "KRAKEN"}
}
\value{
2 lists: one with 7 dataframes, one for each taxonomic level, that contains the relative abundances.
}
\description{
Generates one excel file for each taxonomic level containing the relative abundances of all patients in patient_dir.
}
