% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorFototipo.R
\name{indicadorFototipo}
\alias{indicadorFototipo}
\title{Indicador Fototipo}
\usage{
indicadorFototipo(id, MetadataB)
}
\arguments{
\item{id}{patient's id}

\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}
}
\value{
list with estados_max (drinking status predicted), mensaje_alcohol (message with the prediction), barra_plot (plot that will be include in the report).
}
\description{
It indicates if the skin shows damage because of alcohol drinking by analyzing the relative abundance of determined species.
}
