% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorHabitos.R
\name{indicadorHabitos}
\alias{indicadorHabitos}
\title{Indicador Habitos}
\usage{
indicadorHabitos(id, MetadataB)
}
\arguments{
\item{id}{patient's id}

\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}
}
\value{
list with one message for each habits resolution.
}
\description{
It generates a prediction on smoking, alcohol drinking, making excercise and using cosmetics by analizing the relative abundance of determined subspecies.
}
