% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorTipoPiel.R
\name{indicadorTipoPiel}
\alias{indicadorTipoPiel}
\title{Indicador Tipo de Piel}
\usage{
indicadorTipoPiel(patient_dir, MetadataB)
}
\arguments{
\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}

\item{id}{patient's id}
}
\value{
list with pieles_max (type of skin predicted), mensaje_piel (message with the prediction), barra_plot (plot that will be include in the report).
}
\description{
It indicates if the id's skin is oil, dry or mixed by analyzing the relative abundance of determined species.
}
