% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineGrupal.R
\name{ProcessFirst}
\alias{ProcessFirst}
\title{Run DRAGEN}
\usage{
ProcessFirst(patients_dir, project_id, bs_path)
}
\arguments{
\item{project_id}{it kvsbdkj}

\item{bs_path}{path to executable Base Space}

\item{patient_dir}{fskbfjsf}
}
\description{
kvjdfnkvjdf
}
\examples{
tabla_otus_kraken <- generateOTUsTable_Individual(patient_dir ="~/Biota/Nuevas13/Muestras/33", source= "KRAKEN" )
}
