% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCcontrol.R
\name{QCcontrol}
\alias{QCcontrol}
\title{QC control}
\usage{
QCcontrol(
  patients_dir,
  de_host,
  generate_QCReport_Individual = FALSE,
  generate_QCReport_Grupal = TRUE,
  FastQC_trimmed = FALSE
)
}
\arguments{
\item{patients_dir}{path that points to the directory where all the folders of the samples are stored.}

\item{de_host}{Indicated the software that will be implemented to perform the de-hosting. It can be "Bowtie", "BWA", "RSubread" or "" (none de-hosting).}

\item{generate_QCReport_Individual}{if it is set to TRUE a QC report will be generated for each of the samples in patients_dir.}

\item{generate_QCReport_Grupal}{if it is set to TRUE a QC report will be generated for all the samples in patients_dir.}

\item{FastQC_trimmed}{if it is set to TRUE the QC report will be generated using QC metrics from trimmed FASTQ files instead of the raw FASTQ files.}
}
\value{
qc_metrics is  dataframe that records qc mean, \%Q30, number of sequences, percentage of CG, sequnece length and mapped metrics.
}
\description{
It generates the following tasks:
\itemize{
\item Quality control with FastQC
\item Filter and trimming with Trim Galore
\item De-hosting
\item QC report (individual and grupal).
}
}
