% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_Tax.R
\name{reportCountsTax}
\alias{reportCountsTax}
\title{Report Counts Taxonomy}
\usage{
reportCountsTax(id, MetadataB, patients_dir, de_host, conEukaryota)
}
\arguments{
\item{patient_dir}{is a dataframe fskbfjsf}

\item{source}{kjsnfn}
}
\value{
cants_original dataframe that contains how many subjects from the same taxonomic level were found.
}
\description{
kvjdfnkvjdf
}
\examples{
MetadataB <- as.data.frame(read_excel("~/Daniela/Biota/PipelineBiota/data/Metadata-soloColumnasUsables.xlsx"))
reportCountsTax(id = "37", MetadataB = MetadataB, patients_dir = patients_dir)
}
