% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportMasAbundantes.R
\name{reportMasAbundantes}
\alias{reportMasAbundantes}
\title{Report Most Abundant Bacteria}
\usage{
reportMasAbundantes(patient_dir, MetadataB)
}
\arguments{
\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}

\item{id}{patient's id}
}
\value{
list_plots list with 2 pie charts for every taxonomic level (one that represents the id's composition and the expected composition for its age range)
}
\description{
It describes the relative abundance of the most abundant microorganisms (levels = phylum, genus, species, subspecies).
}
\examples{
reportMasAbundantes(id = "108", MetadataB = MetadataB)
}
