% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDRAGEN.R
\name{RunDRAGEN}
\alias{RunDRAGEN}
\title{Run DRAGEN}
\usage{
RunDRAGEN(
  patient_dir,
  bs_path,
  reference = "hg38",
  de_host,
  conEukaryota = TRUE
)
}
\arguments{
\item{patient_dir}{fskbfjsf}

\item{bs_path}{path to executable Base Space}

\item{project_id}{it kvsbdkj}
}
\description{
kvjdfnkvjdf
}
\examples{
tabla_otus_kraken <- generateOTUsTable_Individual(patient_dir ="~/Biota/Nuevas13/Muestras/33", source= "KRAKEN" )
}
