% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunHuman.R
\name{RunHuman}
\alias{RunHuman}
\title{Run HUMAnN}
\usage{
RunHuman(patient_dir, de_host)
}
\arguments{
\item{patient_dir}{path to the folder of the patient that will be analyzed.}

\item{de_host}{indicates which alignment-based de-hosting software will be used.}
}
\description{
Executes concat of R1 and R2 fastq files, executes HUMAnN 3.0 and generates CPM.
}
