% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunKRAKEN.R
\name{RunKRAKEN}
\alias{RunKRAKEN}
\title{Run KRAKEN 2}
\usage{
RunKRAKEN(patients_dir, confidence = 0.001, de_host)
}
\arguments{
\item{patients_dir}{fskbfjsf}

\item{confidence}{is a score}

\item{path_kraken2}{path to the executable kraken2}
}
\description{
kvjdfnkvjdf
}
\examples{
RunKRAKEN(patients_dir = "/media/16TBDisk/Daniela/Biota/Muestras2daTanda/5/concat5_04/yoconmuestraheritas")
}
