% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLanes.R
\name{splitLanes}
\alias{splitLanes}
\title{Split Lanes}
\usage{
splitLanes(patients_dir)
}
\arguments{
\item{patients_dir}{fskbfjsf}
}
\description{
kvjdfnkvjdf
}
\examples{
splitLanes(patients_dir = "/media/16TBDisk/Daniela/Biota/Muestras2daTanda/5/concat5_04/yoconmuestraheritas")
}
