% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadSamples.R
\name{uploadFolderDrive}
\alias{uploadFolderDrive}
\title{Uploades fastq R1 and de-hosted fastq R1 from server to Drive}
\usage{
uploadFolderDrive(local_folder, drive_folder_url)
}
\arguments{
\item{local_folder}{is the path where all the folders will be taken from.}

\item{folder_url}{is a url.}
}
\description{
Uploads samples from a folder in googledrive and saves them respecting the structure needed.
}
\examples{
drive_folder_url <- "https://drive.google.com/drive/folders/1vcYZAqK74kxEZcZEe74bOQ_f9Hfu9nVW?usp=drive_link"
}
