% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDRAGEN.R
\name{uploadtoBS}
\alias{uploadtoBS}
\title{upload to BS (Base Space)}
\usage{
uploadtoBS(patient_dir, bs_path, de_host = "")
}
\arguments{
\item{patient_dir}{fskbfjsf allows fastq and fastq.gz}

\item{bs_path}{path to base space executable}

\item{project_id}{it kvsbdkj}
}
\description{
kvjdfnkvjdf
}
\examples{
tabla_otus_kraken <- generateOTUsTable_Individual(patient_dir ="~/Biota/Nuevas13/Muestras/33", source= "KRAKEN" )
}
