% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorFagos.R
\name{indicadorFagos}
\alias{indicadorFagos}
\title{Indicador Fagos}
\usage{
indicadorFagos(patient_dir, MetadataB)
}
\arguments{
\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}

\item{id}{patient's id}
}
\value{
mensaje_fagos (message with the prediction).
}
\description{
It indicates if the skin shows presence of phages.
}
