% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorRangoEtarioB.R
\name{indicadorRangoEtario}
\alias{indicadorRangoEtario}
\title{Indicador Rango Etario}
\usage{
indicadorRangoEtario(patient_dir, MetadataB)
}
\arguments{
\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}

\item{id}{patient's id}
}
\value{
list with df_rango_real (dataframe with comparison of relative abundances within the biological age),
rango_final (prediction of age), plot1 (pie chart), combined_plot (combination of pie charts, one for each age range),
mensaje (message with the resolution), barra_plot (plot that will be included in the report).
}
\description{
kvjdfnkvjdf
}
