% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadorVirus.R
\name{indicadorVirus}
\alias{indicadorVirus}
\title{Indicador Virus}
\usage{
indicadorVirus(id, MetadataB)
}
\arguments{
\item{id}{patient's id}

\item{MetadataB}{dataframe that contains the metadata information. id must be within Metadata$ID.}
}
\value{
mensaje_virus (message with the conclusion).
}
\description{
It indicates if the skin present relative abundance of determined viruses above the normal values.
}
