% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPipelineIndividual.R
\name{RunPipelineIndividual}
\alias{RunPipelineIndividual}
\title{Run PipelineBiota Individual}
\usage{
RunPipelineIndividual(patient_dir, runDRAGEN = FALSE)
}
\arguments{
\item{patient_dir}{path to the directory of the patient you want to analyze. It must contain 2 files: ID_R1_fastq and ID_R2_fastq.}

\item{runDRAGEN}{is logical. If it set to TRUE the analyzes will be generated with both KRAKEN and DRAGEN.}
}
\value{
path for 2 reports
}
\description{
Process de patient given from the raw fastq file until 2 reports with the results.
}
\examples{
RunPipelineIndividual(patient_dir = "/media/4tb2/Daniela/Biota/Muestras/180")
}
